refactor(cell/content): split executor.rs by Action family (#244)
executor.rs was 1473 lines (2x the 700-line cap), the worst offender on
the over-cap list in #244. Split the per-action match into one handler
module per behavior family; orchestration stays in executor/mod.rs as a
thin dispatch loop.

  executor/mod.rs    — execute_actions match (orchestration only)
  executor/mission.rs — accept/advance, complete, advance_step,
                        abandon, complete_objective, mission_repeats helper
  executor/inventory.rs — grant (with bandolier seeding + ammo stat),
                          remove (by-instance vs by-type fork),
                          item_container, weapon_stats helpers
  executor/dialog.rs    — display, add_dialog_set, remove_dialog_set,
                          add_dialog, send_interaction_update_if_visible
  executor/stats.rs     — change_stat
  executor/world.rs     — set_interaction_type, set_aggression,
                          destroy_tagged_entity, generate_threat,
                          set_visible, move_waypoint
  executor/counter.rs   — increment, reset
  executor/transport.rs — trigger_transporter, teleport
  executor/tests.rs     — existing test suite, unchanged behavior

The PlaySequence / StartMinigame / SystemMessage / SendMessage /
SetActiveSlot / TriggerChain / wildcard arms are short and don't share
helpers, so they stay inline in mod.rs.

Public surface unchanged: callers (event_dispatch.rs and the parent
module's tests) still import via `executor::execute_actions` and
`super::executor::item_container` respectively. The latter is preserved
through a `#[cfg(test)]`-gated `pub(super) use` re-export.

All 9 executor tests + 38 sibling tests in `cell::content::*` still pass;
clippy clean; cargo fmt applied.
